Add prefix to custom attributes (#1)
This is a constant problem in Android. Attributes do not have name spaces. Attributes of the same name must have the same type. To make this library bullet-proof and to avoid possible conflicts with other libraries `sb_` and `sbg_` prefixes were added. That approach was used in the original library.

Use original library prefix

Co-authored-by: Artem Kuzmyk <artem.kuzmyk@mercedes-benz.io>